+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### 1.1.3 (Unreleased)
 
+- [#431](https://github.com/influxdata/clockface/pull/431): Add `VisibilityInput` component
 - [#428](https://github.com/influxdata/clockface/pull/428): Add missing export for `LinkButton`
 
 #### 1.1.2